+++ Assets/LDtkUnity/Editor/Builders/LDtkBuilderTileset.cs
@@ -73,12 +73,6 @@ namespace LDtkUnity.Editor
                 AddLayerOffset(tilemap);
                 tilemap.SetOpacity(Layer);
             }
-            
-            //add postprocessing for tile layers
-            LDtkPostProcessorCache.AddPostProcessAction(() =>
-            {
-                LDtkPostProcessorInvoker.PostProcessTilesetLayer(LayerGameObject, Layer, _tilesetProvider.Tilemaps.ToArray());
-            });
         }
 
         private void LogPotentialTextureProblems(Texture2D tex)