+++ Assets/LDtkUnity/Editor/Builders/LDtkLevelBackgroundBuilder.cs
@@ -48,11 +48,6 @@ namespace LDtkUnity.Editor
             renderer.sortingOrder = _layerSortingOrder.SortingOrderValue;
 
             ManipulateColorTransform(renderer.transform);
-            
-            LDtkPostProcessorCache.AddPostProcessAction(() =>
-            {
-                LDtkPostProcessorInvoker.PostProcessBackgroundColor(renderer.gameObject);
-            });
         }
 
         private void BuildBackgroundTexture()
@@ -86,11 +81,6 @@ namespace LDtkUnity.Editor
             ManipulateImageTransform(renderer.transform);
 
             _importer.AddBackgroundArtifact(sprite);
-            
-            LDtkPostProcessorCache.AddPostProcessAction(() =>
-            {
-                LDtkPostProcessorInvoker.PostProcessBackgroundTexture(renderer.gameObject);
-            });
         }
 
         private void ManipulateColorTransform(Transform trans)