+++ Assets/LDtkUnity/Runtime/Tools/Const/LDtkAddComponentMenu.cs
@@ -1,7 +1,0 @@
-﻿namespace LDtkUnity
-{
-    internal static class LDtkAddComponentMenu
-    {
-        public const string ROOT = "LDtk/LDtk ";
-    }
-}